feat(api): reject if not owned

+++ src/routes/assignment.rs
@@ -8,21 +8,27 @@ use axum::{
 use hyper::StatusCode;
 
 use crate::{
-    dto::{AssignmentRequest, Claims},
-    entities::assignment::Assignment,
+    dto::{AssignmentRequest, UserInfo},
+    entities::{assignment::Assignment, cohort::Cohort},
     response::Response,
     startup::ApiState,
     ApiError,
 };
 
 pub async fn get_assignment(
-    _: Claims,
+    user: UserInfo,
     State(state): State<Arc<ApiState>>,
     Path(assignment_id): Path<uuid::Uuid>,
 ) -> Result<impl IntoResponse, ApiError> {
-    let assignments = Assignment::find_one(&state.db_pool, assignment_id).await?;
+    let assignment = Assignment::find_one(&state.db_pool, assignment_id).await?;
+    let cohort = Cohort::find_one(&state.db_pool, assignment.cohort_id).await?;
+    if cohort.email != user.email {
+        return Err(ApiError::AuthorizationError(
+            "You are not authorized to view this assignment".to_string(),
+        ));
+    }
     Ok(Response::new(
-        assignments,
+        assignment,
         "Assignment retrieved successfully".to_string(),
         vec![],
     )
@@ -30,10 +36,16 @@ pub async fn get_assignment(
 }
 
 pub async fn create_assignment(
-    _: Claims,
+    user: UserInfo,
     State(state): State<Arc<ApiState>>,
     Json(assignment): Json<AssignmentRequest>,
 ) -> Result<impl IntoResponse, ApiError> {
+    let cohort = Cohort::find_one(&state.db_pool, assignment.cohort_id).await?;
+    if cohort.email != user.email {
+        return Err(ApiError::AuthorizationError(
+            "You are not authorized to create a new assignment in this cohort".to_string(),
+        ));
+    }
     let assignment = Assignment::create(&state.db_pool, assignment).await?;
     Ok(Response::new(
         assignment,
@@ -44,13 +56,19 @@ pub async fn create_assignment(
 }
 
 pub async fn update_assignment(
-    _: Claims,
+    user: UserInfo,
     State(state): State<Arc<ApiState>>,
     Path(assignment_id): Path<uuid::Uuid>,
-    Json(assignment): Json<AssignmentRequest>,
+    Json(input): Json<AssignmentRequest>,
 ) -> Result<impl IntoResponse, ApiError> {
-    Assignment::find_one(&state.db_pool, assignment_id).await?;
-    let assignment = Assignment::update(&state.db_pool, assignment_id, assignment).await?;
+    let assignment = Assignment::find_one(&state.db_pool, assignment_id).await?;
+    let cohort = Cohort::find_one(&state.db_pool, assignment.cohort_id).await?;
+    if cohort.email != user.email {
+        return Err(ApiError::AuthorizationError(
+            "You are not authorized to update this assignment".to_string(),
+        ));
+    }
+    let assignment = Assignment::update(&state.db_pool, assignment_id, input).await?;
     Ok(Response::new(
         assignment,
         "Assignment updated successfully".to_string(),
@@ -60,11 +78,17 @@ pub async fn update_assignment(
 }
 
 pub async fn delete_assignment(
-    _: Claims,
+    user: UserInfo,
     State(state): State<Arc<ApiState>>,
     Path(assignment_id): Path<uuid::Uuid>,
 ) -> Result<impl IntoResponse, ApiError> {
-    Assignment::find_one(&state.db_pool, assignment_id).await?;
+    let assignment = Assignment::find_one(&state.db_pool, assignment_id).await?;
+    let cohort = Cohort::find_one(&state.db_pool, assignment.cohort_id).await?;
+    if cohort.email != user.email {
+        return Err(ApiError::AuthorizationError(
+            "You are not authorized to delete this assignment".to_string(),
+        ));
+    }
     Assignment::delete(&state.db_pool, assignment_id).await?;
     Ok(
         Response::new((), "Assignment deleted successfully".to_string(), vec![])
